+++ src/app/app-routing.module.ts
@@ -96,10 +96,10 @@ const routes: Routes = [
   },
 
   {
-    path: 'joblist',
+    path: 'job-list',
     data: { title: '求人一覧' },
     loadChildren: () =>
-      import('./joblist/joblist.module').then(m => m.JoblistModule),
+      import('./job-list/job-list.module').then(m => m.JoblistModule),
     canLoad: [AuthGuard],
     canActivate: [AuthGuard]
   },
