+++ src/app/app.component.html
@@ -6,7 +6,7 @@
         <mat-list-item routerLink="/">
           ホーム
         </mat-list-item>
-        <mat-list-item routerLink="/joblist">
+        <mat-list-item routerLink="/job-list">
           求人一覧
         </mat-list-item>
       </mat-nav-list>
