+++ src/app/footer/footer.component.html
@@ -27,7 +27,7 @@
       </a>
     </div>
     <div class="footer-box__icon">
-      <a routerLink="/joblist">
+      <a routerLink="/job-list">
         <mat-icon aria-hidden="false">work</mat-icon>
         <span>求人一覧</span>
       </a>
