+++ src/app/home/top/top.component.html
@@ -2,7 +2,7 @@
   <section class="attention">
     <h1>注目の求人</h1>
     <app-attention></app-attention>
-    <button routerLink="/joblist" mat-stroked-button color="primary">
+    <button routerLink="/job-list" mat-stroked-button color="primary">
       求人一覧へ >>
     </button>
   </section>
