+++ src/app/job-list/job-list-routing.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { JoblistComponent } from './joblist/joblist.component';
-
+import { JoblistComponent } from './job-list/job-list.component';
 
 const routes: Routes = [
   {
@@ -14,4 +13,4 @@ const routes: Routes = [
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
-export class JoblistRoutingModule { }
+export class JoblistRoutingModule {}
