+++ src/app/job-list/job-list/job-list.component.scss
