+++ src/app/job-list/job-list/job-list.component.ts
@@ -5,8 +5,8 @@ import { JobPostService } from 'src/app/services/job-post.service';
 
 @Component({
   selector: 'app-joblist',
-  templateUrl: './joblist.component.html',
-  styleUrls: ['./joblist.component.scss']
+  templateUrl: './job-list.component.html',
+  styleUrls: ['./job-list.component.scss']
 })
 export class JoblistComponent implements OnInit {
   jobs$: Observable<DetailJob[]> = this.jobPostService.getAllJob();
